+++ source/test/LoadBalanceCorrectness/main-lbcorrectness.C
@@ -1594,7 +1594,7 @@ void enforceNesting(
 
    const hier::BoxLevel &L0 = L1_to_L0.getHead();
 
-   const int cell_count = L1.getGlobalNumberOfCells();
+   const long int cell_count = L1.getGlobalNumberOfCells();
 
    /*
     * Make L1 nest inside L0 by nesting_width.
